# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TKK Control API",
        "description": "API documentation for TKK-Control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
        "/api/checks": {
            "post": {
                "tags": [
                    "Checks"
                ],
                "summary": "Register a check or dropout at a control",
                "description": "Creates a record of passage (check) or dropout for a participant at a specific control.",
                "operationId": "36827b3a46d057db7a74ed7237d2bec8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "control_id",
                                    "participant_id",
                                    "type"
                                ],
                                "properties": {
                                    "control_id": {
                                        "description": "ID of the control",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "participant_id": {
                                        "description": "ID of the participant",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "type": {
                                        "description": "Type of record",
                                        "type": "string",
                                        "enum": [
                                            "check",
                                            "abandon"
                                        ],
                                        "example": "check"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checked successfully"
                                        },
                                        "check": {
                                            "description": "Created record data",
                                            "type": "object"
                                        },
                                        "control": {
                                            "description": "Updated control state",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or duplicate record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "This participant already has a 'check' record at this control."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Control or participant not found"
                    }
                }
            }
        },
        "/api/controls/{control}/checks": {
            "get": {
                "tags": [
                    "Checks"
                ],
                "summary": "Get participants who checked at a control",
                "description": "Returns the list of participants who have passed a given control, ordered by timestamp descending.",
                "operationId": "22cf1c42f426f6864e9c85aaf09909cb",
                "parameters": [
                    {
                        "name": "control",
                        "in": "path",
                        "description": "ID of the control",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of participants who passed the control",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "dorsal": {
                                                "type": "integer",
                                                "example": 23
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Maria López"
                                            },
                                            "checked_at": {
                                                "type": "string",
                                                "example": "2025-09-21 09:30:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}